+++ tests/index.ts
@@ -4,6 +4,7 @@ import 'reflect-metadata';
 
 // tests
 import './angular2Apollo';
+import './utils';
 import './apolloQueryObservable';
 import './apolloQueryPipe';
-import './apolloDecorator/index';
+import './apolloDecorator';